move QModule to its proper place
