+++ all/approximation/q_network.py
@@ -1,4 +1,5 @@
-from all.nn import QModule
+import torch
+from all.nn import RLNetwork
 from .approximation import Approximation
 
 class QNetwork(Approximation):
@@ -6,14 +7,22 @@ class QNetwork(Approximation):
             self,
             model,
             optimizer,
-            num_actions,
             name='q',
             **kwargs
     ):
-        model = QModule(model, num_actions)
+        model = QModule(model)
         super().__init__(
             model,
             optimizer,
             name=name,
             **kwargs
         )
+
+class QModule(RLNetwork):
+    def forward(self, states, actions=None):
+        values = super().forward(states)
+        if actions is None:
+            return values
+        if isinstance(actions, list):
+            actions = torch.tensor(actions, device=self.device)
+        return values.gather(1, actions.view(-1, 1)).squeeze(1)
