+++ all/nn/__init__.py
@@ -214,21 +214,6 @@ class TanhActionBound(nn.Module):
         return torch.tanh(x) * self.weight + self.bias
 
 
-class QModule(nn.Module):
-    def __init__(self, model, num_actions):
-        super().__init__()
-        self.device = next(model.parameters()).device
-        self.model = RLNetwork(model, (num_actions,))
-
-    def forward(self, states, actions=None):
-        values = self.model(states)
-        if actions is None:
-            return values
-        if isinstance(actions, list):
-            actions = torch.tensor(actions, device=self.device)
-        return values.gather(1, actions.view(-1, 1)).squeeze(1)
-
-
 class VModule(nn.Module):
     def __init__(self, model):
         super().__init__()
